+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 
 - Support for mount filesystems
 - Check staging and publishing targets
+- Exit on binding error
 
 
 0.0.2 (2018-06-19)